+++ src/main/java/br/com/brenohff/later/api/EventController.java
@@ -1,6 +1,6 @@
-package br.com.brenohff.later.controller;
+package br.com.brenohff.later.api;
 
-import br.com.brenohff.later.models.LTEvent;
+import br.com.brenohff.later.model.LTEvent;
 import br.com.brenohff.later.service.EventService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -18,33 +18,33 @@ public class EventController {
     @Autowired
     EventService service;
 
-    @RequestMapping(value = "/saveEvent", method = RequestMethod.POST)
+    @PostMapping(value = "/saveEvent")
     public ResponseEntity<Void> saveEvent(@RequestPart("event") String event, @RequestPart MultipartFile file) {
         service.saveEvent(event, file);
         return ResponseEntity.status(HttpStatus.OK).build();
     }
 
-    @RequestMapping(value = "/getAll")
+    @GetMapping(value = "/getAll")
     public List<LTEvent> getAllEvents() {
         return service.getAllEvents();
     }
 
-    @RequestMapping(value = "/getPublic")
+    @GetMapping(value = "/getPublic")
     public ResponseEntity<List<LTEvent>> getPublic() {
         return ResponseEntity.status(HttpStatus.OK).body(service.getPublic());
     }
 
-    @RequestMapping(value = "/getEventById")
+    @GetMapping(value = "/getEventById")
     public LTEvent getEventById(@RequestParam(value = "event_id") Long event_id) {
         return service.getEventById(event_id);
     }
 
-    @RequestMapping(value = "/getEventsByUser")
+    @GetMapping(value = "/getEventsByUser")
     public List<LTEvent> getEventsByUser(@RequestParam(value = "user_id") String user_id) {
         return service.getEventsByUser(user_id);
     }
 
-    @RequestMapping(value = "/delete")
+    @GetMapping(value = "/delete")
     public void delete(@RequestParam(value = "event_id") Long event_id) {
         service.delete(event_id);
     }
