+++ src/main/java/br/com/brenohff/later/repository/CategoryEventRepository.java
@@ -2,8 +2,8 @@ package br.com.brenohff.later.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import br.com.brenohff.later.models.LTCategoryEvent;
-import br.com.brenohff.later.models.LTCategoryEventPK;
+import br.com.brenohff.later.model.LTCategoryEvent;
+import br.com.brenohff.later.model.LTCategoryEventPK;
 
 public interface CategoryEventRepository extends JpaRepository<LTCategoryEvent, LTCategoryEventPK>{
