+++ src/main/java/br/com/brenohff/later/repository/CategoryRepository.java
@@ -2,7 +2,7 @@ package br.com.brenohff.later.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import br.com.brenohff.later.models.LTCategory;
+import br.com.brenohff.later.model.LTCategory;
 
 public interface CategoryRepository extends JpaRepository<LTCategory, Long>{
