+++ src/main/java/br/com/brenohff/later/repository/EventRepository.java
@@ -6,9 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import br.com.brenohff.later.models.LTEvent;
-
-import javax.websocket.server.PathParam;
+import br.com.brenohff.later.model.LTEvent;
 
 public interface EventRepository extends JpaRepository<LTEvent, Long> {
