+++ src/main/java/br/com/brenohff/later/service/ChatService.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import br.com.brenohff.later.models.LTChat;
+import br.com.brenohff.later.model.LTChat;
 import br.com.brenohff.later.repository.ChatRepository;
 
 @Service
