+++ src/main/java/br/com/brenohff/later/service/UserService.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import br.com.brenohff.later.models.LTUser;
+import br.com.brenohff.later.model.LTUser;
 import br.com.brenohff.later.repository.UserRepository;
 import br.com.brenohff.later.service.exceptions.ObjectAlreadyExists;
 import br.com.brenohff.later.service.exceptions.ObjectNotFound;
