+++ src/main/resources/application.properties
@@ -6,12 +6,12 @@ spring.jpa.database-platform=org.hibernate.dialect.MySQL5Dialect
 
 
 ################################## PORT #################################
-server.port=5000
+server.port=8080
 
 
 ################################## H2 ##################################
-#spring.h2.console.enabled=true
-#spring.h2.console.path=/h2-console
+spring.h2.console.enabled=true
+spring.h2.console.path=/h2-console
 
 
 ################################## LOCALHOST ##################################
